Added Content to 2014 Repo Herve Devos
Added folder structure: botball2013Routines which contains the code
from last year, also created a botball2014 folder that should contain
the work for this year with separete sub folders for infodoc,
KippyRoutines for KIPR and createRoutines for create so we can keep
things organized.
Added one program in KippyRoutines that attempts to move KIPR in inches
and support right and left right angle turns. It may have to be tweaked
and needs to be retested when the units are shipped back. Need to
verify whether battery levels will impact the constants as well.